set external input value to old 0.5, vs recent 0.1, deterministic.
I had reasoned that external-input-wt
s/b <= 0.25?, i.e. s.t. 2 * neg-weight + 3 * 2/3 * pos-weight <= 1
but in fact this means that the net fails to settle too often.
Changing this to 0.5 always settles, it seems, though sometimes it
takes a while.

I also got rid of the randomizer to this value.  That was intended
to break out of paradoxical states with the old small values for
external-input-wt, but it wasn't always successful, even with the
wt equal to 0.25.  0.5 alone, without randomization, is enough.